Removed some stl, changed the way logging works, removed old logging functions.

+++ code/external/graphics_api/graphics_api/ogl/ogl_vertex_format.hpp
@@ -42,7 +42,7 @@ struct Attribute
   char          name[32];
   GLenum        type    = GL_NONE;
   uint32_t      size    = 0;
-  uint32_t      pointer = 0;
+  uintptr_t     pointer = 0;
 };
